Removed use of <functional> std header

+++ src/LocalBinaryPattern.cpp
@@ -144,7 +144,7 @@ cv::Mat_<float> LocalBinaryPattern::extract( const cv::Rect& rct) const
     const cv::Vec3b fv = extractBinaryPattern(rct); // 3-bytes
     const int v = static_cast<int>(fv[0]) << 16 | static_cast<int>(fv[1]) << 8 | static_cast<int>(fv[2]);
     cv::Mat_<float> fvm(1,1);
-    *fvm.ptr<float>() = v;
+    *fvm.ptr<float>() = static_cast<float>(v);
     return fvm;
 }   // end extract
